backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
aiosqlite==0.20.0
httpx==0.28.1
python-dotenv==1.1.0
python-multipart==0.0.18
pdfplumber==0.11.4
Pillow==11.1.0
